server/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.15.1,<0.16
django-cors-headers>=4.3.1,<4.4
Pillow>=10.2.0,<10.3